refractor: TemplateCreatorPage -> move renderer useeffect to Templaterenderer component
